Conflicts:
	novabill-android-client/AndroidManifest.xml
	novabill-android-client/pom.xml
	novabill-android-shared/pom.xml
	novabill-gwt-client/pom.xml
	novabill-gwt-shared/pom.xml
	novabill-scala-utils/pom.xml
	novabill-server/pom.xml
	novabill-server/src/main/resources/reports/tidy/tidyCredNote.jrxml
	novabill-server/src/main/resources/reports/tidy/tidyEstimation.jrxml
	novabill-server/src/main/resources/reports/tidy/tidyInvoice.jrxml
	novabill-server/src/main/resources/reports/tidy/tidyTransDoc.jrxml
	pom.xml

+++ novabill-gwt-client/src/main/java/com/novadart/novabill/frontend/client/view/center/ItemTable.java
@@ -29,15 +29,31 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 	public static interface Handler{
 		public void onDelete(AccountingDocumentItemDTO item);
 		public void onUpdate(AccountingDocumentItemDTO item);
+		public void onMoveUp(AccountingDocumentItemDTO value);
+		public void onMoveDown(AccountingDocumentItemDTO value);
 	}
 
 	private final Handler handler;
+	private boolean locked = false;
 
+	private final MoveUpDownCell moveUpDownCell;
 
-	public ItemTable(Handler handler) {
-		super(2000);
+	public ItemTable(boolean displayWeight, Handler handler) {
+		super(10000);
 
 		this.handler = handler;
+
+
+		moveUpDownCell = new MoveUpDownCell(this.handler);
+		final Column<AccountingDocumentItemDTO, AccountingDocumentItemDTO> moveUpDown = 
+				new Column<AccountingDocumentItemDTO, AccountingDocumentItemDTO>(moveUpDownCell) {
+
+			@Override
+			public AccountingDocumentItemDTO getValue(AccountingDocumentItemDTO object) {
+				return object;
+			}
+		};
+		addColumn(moveUpDown);
 
 
 		//Sku
@@ -69,6 +85,12 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void update(int index, AccountingDocumentItemDTO object, String value) {
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					descEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
 				if(Validation.isEmpty(value)){
 					Notification.showMessage(I18N.INSTANCE.errorClientData());
 				} else if(!Validation.isWithinSize(value, DESCRIPTION_MAX_LENGTH)){
@@ -85,6 +107,44 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 			}
 		});
 		addColumn(nameDescription, I18N.INSTANCE.nameDescription());
+
+		//unity of measure
+		final EditTextCell unitEditCell = new EditTextCell();
+		Column<AccountingDocumentItemDTO, String> unitOfMeasure =
+				new Column<AccountingDocumentItemDTO, String>(unitEditCell) {
+
+			@Override
+			public String getValue(AccountingDocumentItemDTO object) {
+				return object.getUnitOfMeasure();
+			}
+		};
+		unitOfMeasure.setFieldUpdater(new FieldUpdater<AccountingDocumentItemDTO, String>() {
+
+			@Override
+			public void update(int index, AccountingDocumentItemDTO object, String value) {
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					unitEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
+
+				if(!Validation.isWithinSize(value, UNIT_OF_MEASURE_MAX_SIZE)){
+					Notification.showMessage(I18NM.get.textLengthError(UNIT_OF_MEASURE_MAX_SIZE));
+				} else {
+					object.setUnitOfMeasure(value);
+					ItemTable.this.handler.onUpdate(object);
+				}
+
+				unitEditCell.clearViewData(object);
+				redrawRow(index);
+			}
+		});
+		SafeHtmlBuilder shb = new SafeHtmlBuilder();
+		shb.appendHtmlConstant("<span title=\""+I18N.INSTANCE.unityOfMeasureExtended()+"\">");
+		shb.appendEscaped(I18N.INSTANCE.unityOfMeasure());
+		shb.appendHtmlConstant("</span>");
+		addColumn(unitOfMeasure, shb.toSafeHtml());
 
 
 		//quantity
@@ -105,19 +165,24 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void update(int index, AccountingDocumentItemDTO object, String value) {
-				if(object.getQuantity() != null){
-					try{
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					qtyEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
 
-						BigDecimal newQty = CalcUtils.parseValue(value);
-						object.setQuantity(newQty);
-						ItemTable.this.handler.onUpdate(object);
+				try{
 
-					} catch(NumberFormatException e){
+					BigDecimal newQty = CalcUtils.parseValue(value);
+					object.setQuantity(newQty);
+					ItemTable.this.handler.onUpdate(object);
 
-						Notification.showMessage(I18N.INSTANCE.errorClientData());
+				} catch(NumberFormatException e){
 
-					}
-				}			
+					Notification.showMessage(I18N.INSTANCE.errorClientData());
+
+				}
 
 				qtyEditCell.clearViewData(object);
 				redrawRow(index);
@@ -125,38 +190,55 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 		});
 		addColumn(quantity, I18N.INSTANCE.quantity());
 
+		Column<AccountingDocumentItemDTO, String> weight = null;
+		if(displayWeight) {
 
-		//unity of measure
-		final EditTextCell unitEditCell = new EditTextCell();
-		Column<AccountingDocumentItemDTO, String> unitOfMeasure =
-				new Column<AccountingDocumentItemDTO, String>(unitEditCell) {
+			//weight
+			final EditTextCell weightEditCell = new EditTextCell();
+			weight = new Column<AccountingDocumentItemDTO, String>(weightEditCell) {
 
-			@Override
-			public String getValue(AccountingDocumentItemDTO object) {
-				return object.getUnitOfMeasure();
-			}
-		};
-		unitOfMeasure.setFieldUpdater(new FieldUpdater<AccountingDocumentItemDTO, String>() {
+				@Override
+				public String getValue(AccountingDocumentItemDTO object) {
+					if(object.getWeight() == null) {
+						return "";
+					}
 
-			@Override
-			public void update(int index, AccountingDocumentItemDTO object, String value) {
-				if(!Validation.isWithinSize(value, UNIT_OF_MEASURE_MAX_SIZE)){
-					Notification.showMessage(I18NM.get.textLengthError(UNIT_OF_MEASURE_MAX_SIZE));
-				} else if(object.getUnitOfMeasure() != null){
-					object.setUnitOfMeasure(value);
-					ItemTable.this.handler.onUpdate(object);
+					return NumberFormat.getDecimalFormat().format(object.getWeight());
 				}
+			};
+			weight.setFieldUpdater(new FieldUpdater<AccountingDocumentItemDTO, String>() {
 
-				unitEditCell.clearViewData(object);
-				redrawRow(index);
-			}
-		});
-		SafeHtmlBuilder shb = new SafeHtmlBuilder();
-		shb.appendHtmlConstant("<span title=\""+I18N.INSTANCE.unityOfMeasureExtended()+"\">");
-		shb.appendEscaped(I18N.INSTANCE.unityOfMeasure());
-		shb.appendHtmlConstant("</span>");
-		addColumn(unitOfMeasure, shb.toSafeHtml());
+				@Override
+				public void update(int index, AccountingDocumentItemDTO object, String value) {
+					if(locked){
+						Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+						weightEditCell.clearViewData(object);
+						redrawRow(index);
+						return;
+					}
 
+					try{
+
+						if(value.isEmpty()){
+							object.setDiscount(null);
+						}  else {
+							BigDecimal newWeight = CalcUtils.parseValue(value);
+							object.setWeight(newWeight);
+							ItemTable.this.handler.onUpdate(object);
+						}
+						
+					} catch(NumberFormatException e){
+
+						Notification.showMessage(I18N.INSTANCE.errorClientData());
+
+					}
+
+					weightEditCell.clearViewData(object);
+					redrawRow(index);
+				}
+			});
+			addColumn(weight, I18N.INSTANCE.weight());
+		}
 
 		//VAT
 		final EditTextCell taxEditCell = new EditTextCell();
@@ -176,19 +258,24 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void update(int index, AccountingDocumentItemDTO object, String value) {
-				if(object.getTax() != null){
-					try{
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					taxEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
 
-						BigDecimal newTax = CalcUtils.parseValue(value);
-						object.setTax(newTax);
-						ItemTable.this.handler.onUpdate(object);
+				try{
 
-					} catch(NumberFormatException e){
+					BigDecimal newTax = CalcUtils.parseValue(value);
+					object.setTax(newTax);
+					ItemTable.this.handler.onUpdate(object);
 
-						Notification.showMessage(I18N.INSTANCE.errorClientData());
+				} catch(NumberFormatException e){
 
-					}
-				}			
+					Notification.showMessage(I18N.INSTANCE.errorClientData());
+
+				}
 
 				taxEditCell.clearViewData(object);
 				redrawRow(index);
@@ -215,18 +302,23 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void update(int index, AccountingDocumentItemDTO object, String value) {
-				if(object.getPrice() != null){
-					try{
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					priceEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
 
-						BigDecimal newPrice = CalcUtils.parseCurrency(value);
-						object.setPrice(newPrice);
-						ItemTable.this.handler.onUpdate(object);
+				try{
 
-					} catch(NumberFormatException e){
+					BigDecimal newPrice = CalcUtils.parseCurrency(value);
+					object.setPrice(newPrice);
+					ItemTable.this.handler.onUpdate(object);
 
-						Notification.showMessage(I18N.INSTANCE.errorClientData());
+				} catch(NumberFormatException e){
 
-					}
+					Notification.showMessage(I18N.INSTANCE.errorClientData());
+
 				}
 				priceEditCell.clearViewData(object);
 				redrawRow(index);
@@ -253,21 +345,26 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void update(int index, AccountingDocumentItemDTO object, String value) {
-				if(object.getDiscount() != null){
-					try{
-						if(value.isEmpty()){
-							object.setDiscount(null);
-						} else {
-							BigDecimal newDiscount = CalcUtils.parseValue(value);
-							object.setDiscount(newDiscount);
-						}
-						ItemTable.this.handler.onUpdate(object);
-					} catch(NumberFormatException e){
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					discountEditCell.clearViewData(object);
+					redrawRow(index);
+					return;
+				}
 
-						Notification.showMessage(I18N.INSTANCE.errorClientData());
-
+				try{
+					if(value.isEmpty()){
+						object.setDiscount(null);
+					} else {
+						BigDecimal newDiscount = CalcUtils.parseValue(value);
+						object.setDiscount(newDiscount);
 					}
-				}			
+					ItemTable.this.handler.onUpdate(object);
+				} catch(NumberFormatException e){
+
+					Notification.showMessage(I18N.INSTANCE.errorClientData());
+
+				}
 
 				discountEditCell.clearViewData(object);
 				redrawRow(index);
@@ -296,6 +393,10 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 
 			@Override
 			public void execute(AccountingDocumentItemDTO object) {
+				if(locked){
+					Notification.showMessage(I18N.INSTANCE.lockedItemsTableAlert());
+					return;
+				}
 				ItemTable.this.handler.onDelete(object);
 			}
 		};
@@ -313,8 +414,11 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 		setWidth("99%");
 		setColumnWidth(sku, 5, Unit.PCT);
 		setColumnWidth(nameDescription, 37, Unit.PCT);
-		setColumnWidth(quantity, 6, Unit.PCT);
 		setColumnWidth(unitOfMeasure, 6, Unit.PCT);
+		setColumnWidth(quantity, 6, Unit.PCT);
+		if(displayWeight) {
+			setColumnWidth(weight, 6, Unit.PCT);
+		}
 		setColumnWidth(price, 6, Unit.PCT);
 		setColumnWidth(discount, 6, Unit.PCT);
 		setColumnWidth(tax, 6, Unit.PCT);
@@ -324,4 +428,9 @@ public class ItemTable extends CellTable<AccountingDocumentItemDTO> {
 		setLoadingIndicator(null);
 	}
 
+	public void setLocked(boolean locked) {
+		this.locked = locked;
+		this.moveUpDownCell.setLocked(locked);
+		redraw();
+	}
 }
